% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMR.R
\name{MetaMR_selectionmodel}
\alias{MetaMR_selectionmodel}
\title{MetaMR using a simple random-effects meta-analysis model on a list of pre-selected variants}
\usage{
MetaMR_selectionmodel(
  sumstat_beta_list,
  sumstat_se_list,
  is_overlap = FALSE,
  r_mat_list = NA,
  check_zeros = TRUE,
  zero_thresh = 1e-08,
  set.init.params = FALSE,
  tau_mu_zero = FALSE,
  tau_delta_zero = FALSE,
  select_method = c("single_exposure", "minp", "fisher"),
  single_exp_pop = 1,
  select_pthresh = 0.05,
  resample_number = 1e+05,
  resample_seed = 2025,
  standardized = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sumstat_beta_list}{a list of vectors of GWAS effect size estimates, each with length K+1: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations. Each vector in the list represents summary statistics for one of N variants.}

\item{sumstat_se_list}{a list of vectors of the standard errors for GWAS effect size estimates in sumstat_beta_list}

\item{is_overlap}{A boolean describing whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat_list}{a list of matrices of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap.}

\item{check_zeros}{whether to check if either variance parameter (tau_mu) or (tau_delta) was estimated to be exactly zero, which can destabilize the algorithm. If either estimated variance parameter falls below a certain threshold, we rerun optimization fixing that component at exactly zero.}

\item{zero_thresh}{the threshold specified in check_zeros. Is 1e-8 by default.}

\item{set.init.params}{Whether the optim() function should use the observed data to set initial parameters for optimization.}

\item{tau_mu_zero}{Whether tau_mu should be fixed at zero when performing MetaMR.}

\item{tau_delta_zero}{Whether tau_delta should be fixed at zero when performing MetaMR.}

\item{select_method}{The method by which variants were selected: "single_exposure" = a single population's exposure association p-value, "minp" = minimum exposure association p-value across all populations, "fisher" = Fisher's method}

\item{single_exp_pop}{If select_method = "single_exposure", what exposure population to apply it to}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{resample_number}{If select_method = "fisher", the number of resamples performed to estimate the probability of a variant being selected under Fisher's method.}

\item{resample_seed}{The seed used for resampling}

\item{standardized}{A boolean indicating whether summary statistics are standardized; i.e. whether they have the same set of standard errors. If they do, we only need to calculate the probability of selection once across all variants.}

\item{verbose}{A boolean indicating whether to print out current values of parameters at every iteration of optim(). Use for troubleshooting purposes.}
}
\value{
a list that includes point estimates for the exposure-outcome causal effect \eqn{\gamma} and variance terms \eqn{\tau_{\mu}} and \eqn{\tau_{\delta}} estimated using maximum likelihood, the covariance matrix of these point estimates (calculated using the inverse of the Hessian), and the results of a likelihood ratio test that tests against null hypothesis \eqn{\gamma = 0} (including the test statistic and p-value)
}
\description{
Like MetaMR_simplemodel, but now assumes that the variants under consideration have passed some selection method. Currently supports variants being selected based on the p-value of exposure association in a single population, their minimum p-value of exposure association across all populations, or the p-value of Fisher's method when applied to the exposure summary statistics for that variant. This method currently does NOT check whether variants being considered fulfills variant selection, but this may be included in the future.
}
\examples{
nvar_causal <- 200
gamma <- 0.1
h2 <- 0.001*nvar_causal
persnp_h2 <- h2 / nvar_causal
tau_mu <- 0.9*persnp_h2
tau_delta <- persnp_h2 - tau_mu
N_k <- c(100000, 5000, 50000)
sigma_Xk <- 1/sqrt(N_k)
special_conditions <- "selection_allvars"
SE_list_causal <- rep(list(sigma_Xk), nvar_causal)
nrep <- 200
seed <- 2025
observed_data_causal <- simplemodel_sim(gamma = gamma, tau_mu = tau_mu,
                                       tau_delta = tau_delta,
                                       SE_list = SE_list_causal,
                                       vars = nvar_causal,
                                       pops = 2,
                                       seed = seed)
sumstat_beta_list_causal <- apply(observed_data_causal$beta_matrix, MARGIN = 1,
                                  function(x){return(x)}, simplify = FALSE)
crit.p <- 0.05/nvar_causal
select_minp <- which(abs(observed_data_causal$beta_matrix[,2]/observed_data_causal$se_matrix[,2]) >
                     abs(qnorm(crit.p/2)) |
abs(observed_data_causal$beta_matrix[,3]/observed_data_causal$se_matrix[,3]) >
abs(qnorm(crit.p/2)))
sumstat_beta_list_causal <- apply(observed_data_causal$beta_matrix, MARGIN = 1,
                                  function(x) {return(x)}, simplify = FALSE)
minp_metamr <- MetaMR_selectionmodel(sumstat_beta_list = sumstat_beta_list_causal[select_minp],
                                     sumstat_se_list = SE_list_causal[select_minp],
                                     is_overlap = FALSE, r_mat_list= NA,
                                     check_zeros = TRUE, zero_thresh = 1e-8,
                                     set.init.params = FALSE,
                                     tau_mu_zero = FALSE, tau_delta_zero = FALSE,
                                     select_method = "minp",
                                     single_exp_pop = 1, select_pthresh = 0.05,
                                     resample_number = 1e5, resample_seed = 2025,
                                     standardized = TRUE, verbose = FALSE)
}
