% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_ash.R
\name{metamrash_loglik.single}
\alias{metamrash_loglik.single}
\title{Log-likelihood function incorporating variant selection and ASH pleiotropy (single-variant)}
\usage{
metamrash_loglik.single(
  sumstat_beta,
  sumstat_se,
  is_overlap = FALSE,
  r_mat = NA,
  kernel_matrix = NA,
  gamma,
  tau_mu,
  tau_delta,
  ash_pi = rep(1/7, 7),
  sigma_c = c(0, 10^seq(-10, 0, by = 2)),
  tau_mu_log = FALSE,
  tau_delta_log = FALSE,
  select_method = c("single_exposure", "minp", "fisher", "none"),
  single_exp_pop = 1,
  select_pthresh = 0.05,
  random_seed = 2026,
  mc_iter = 250
)
}
\arguments{
\item{sumstat_beta}{the GWAS effect size estimates for a single variant across K+1 studies: first, for the outcome in the target population, then across K exposure studies.}

\item{sumstat_se}{the standard errors of the effect size estimates in sumstat_beta}

\item{is_overlap}{a boolean whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat}{a matrix of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap.}

\item{kernel_matrix}{a K+1 x K+1 correlation matrix that captures correlations in exposure effect size heterogeneity across all populations (including the target population).}

\item{gamma}{The exposure-outcome causal effect}

\item{tau_mu}{The variance of the normal distribution underlying mu. May be log-transformed.}

\item{tau_delta}{The variance of the normal distributions underlying delta. May be log-transformed.}

\item{ash_pi}{A vector of length c representing the mixture probabilities of each of the components of the ASH distribution we use. By default, we assume c = 7 and that probabilities are equally distributed across all mixture variance parameters (1/7)}

\item{sigma_c}{A vector of length c representing the variance of each of the components of the ASH distribution we used. By default, we specify c = 7 with components 0, 1e-10, 1e-8, 1e-6, 1e-4, 1e-2, and 1.}

\item{tau_mu_log}{Whether tau_mu is log-transformed}

\item{tau_delta_log}{Whether tau_delta is log-transformed}

\item{select_method}{The method by which variants were selected: "single_exposure" = a single population's exposure association p-value, "minp" = minimum exposure association p-value across all populations, "fisher" = Fisher's method.}

\item{single_exp_pop}{If select_method = "single_exposure", what exposure study it applies to. By default, we assume it is the first.}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{random_seed}{For Fisher's method of selection, what random seed to use for sampling of observations}

\item{mc_iter}{For Fisher's method of selection, how many random samples to take}
}
\value{
A log-likelihood for the summary statistics of a single (selected) variant given parameters gamma, mu, tau, ash_pi
}
\description{
A conditional log-likelihood function that models pleiotropy using the adaptive shrinkage (ASH) prior distribution proposed by Stephens et al (2015), accounting for variants being pre-selected using either a single population's exposure association p-value, minimum exposure association p-value across all populations, or Fisher's method. The ASH distribution is a mixture of normals with mean zero, and fixed variance components.
}
