% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_selection.R
\name{numint_prob}
\alias{numint_prob}
\title{Calculating the probability of a variant being selected using numerical integration (multi-exposure minimum p-value)}
\usage{
numint_prob(tau_mu, tau_delta, r_mat, SE_vector, select_pthresh)
}
\arguments{
\item{tau_mu}{The variance of the normal distribution underlying mu.}

\item{tau_delta}{The variance of the normal distributions underlying delta.}

\item{r_mat}{the matrix from r_mat in selection_loglik_single, but for only exposure summary statistics (i.e. first row and column removed)}

\item{SE_vector}{the vector from sumstat_se in selection_loglik_single, but for only exposure summary statistics (i.e. first entry removed)}

\item{select_pthresh}{The p-value threshold for variant selection}
}
\value{
The probability that a variant was selected under a minimum exposure association p-value selection strategy
}
\description{
Calculating the probability of a variant being selected using numerical integration (multi-exposure minimum p-value)
}
