% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_selection.R
\name{resample_prob}
\alias{resample_prob}
\title{Calculating the probability of a variant being selected using resampling (Fisher's method)}
\usage{
resample_prob(
  resample_number,
  resample_seed,
  tau_mu,
  tau_delta,
  r_mat,
  SE_vector,
  select_pthresh
)
}
\arguments{
\item{resample_number}{the number of resamples performed to estimate the probability of a variant being selected under Fisher's method.}

\item{resample_seed}{The seed used for resampling}

\item{tau_mu}{The variance of the normal distribution underlying mu.}

\item{tau_delta}{The variance of the normal distributions underlying delta.}

\item{r_mat}{the matrix from r_mat in selection_loglik_single, but for only exposure summary statistics (i.e. first row and column removed)}

\item{SE_vector}{the vector from sumstat_se in selection_loglik_single, but for only exposure summary statistics (i.e. first entry removed)}

\item{select_pthresh}{The p-value threshold for variant selection}
}
\value{
The probability that a variant was selected under a Fisher's method selection strategy
}
\description{
Calculating the probability of a variant being selected using resampling (Fisher's method)
}
