% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_ash.R
\name{resample_prob_kernel}
\alias{resample_prob_kernel}
\title{Calculating the probability of a variant being selected by Fisher's method of meta-analysis using Monte-Carlo sampling}
\usage{
resample_prob_kernel(
  resample_number,
  resample_seed,
  tau_mu,
  tau_delta,
  r_mat,
  SE_vector,
  select_pthresh,
  kernel_matrix
)
}
\arguments{
\item{resample_number}{the number of resamples performed to estimate the probability of a variant being selected under Fisher's method.}

\item{resample_seed}{The seed used for resampling}

\item{tau_mu}{The variance of the normal distribution underlying mu.}

\item{tau_delta}{The variance of the normal distributions underlying delta.}

\item{r_mat}{the matrix from r_mat in metamrash_loglik.single, but for only exposure summary statistics (i.e. first row and column removed)}

\item{SE_vector}{the vector from sumstat_se in metamrash_loglik.single, but for only exposure summary statistics (i.e. first entry removed)}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{kernel_matrix}{the kernel matrix from kernel_matrix in metamrash_loglik.single, but only for the K exposure studies (i.e. first row and column removed)}
}
\value{
The probability that a variant was selected under Fisher's method for variant selection
}
\description{
Calculating the probability of a variant being selected by Fisher's method of meta-analysis using Monte-Carlo sampling
}
