% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_selection.R
\name{selection_loglik_optimize}
\alias{selection_loglik_optimize}
\title{Optimizing the conditional log-likelihood under variant selection for a list of selected variants.}
\usage{
selection_loglik_optimize(
  sumstat_beta_list,
  sumstat_se_list,
  is_overlap = FALSE,
  r_mat_list = NA,
  is.fixed = c(FALSE, FALSE, FALSE),
  fix.params = c(NA, NA, NA),
  set.init.params = FALSE,
  tau_mu_log = FALSE,
  tau_delta_log = FALSE,
  optim_method = c("Nelder-Mead", "L-BFGS-B", "Brent"),
  select_method = c("single_exposure", "minp", "fisher"),
  single_exp_pop = 1,
  select_pthresh = 0.05,
  resample_number = 1e+05,
  resample_seed = 2025,
  standardized = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sumstat_beta_list}{a list of vectors of GWAS effect size estimates with length K+1: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations. Each vector in the list represents summary statistics for one of many variants.}

\item{sumstat_se_list}{a list of vectors of the standard errors for GWAS effect size estimates in sumstat_beta_list}

\item{is_overlap}{a boolean describing whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat_list}{a list of matrices of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap.}

\item{is.fixed}{which parameters of interested should be fixed during optimization. Follows order (gamma, tau_mu, tau_delta).}

\item{fix.params}{if a parameter is set as fixed in is.fixed, what it should be. Follows order (gamma, tau_mu, tau_delta). Our function will check to make sure is.fixed and fix.params are compatible with each other.}

\item{set.init.params}{Whether to manually set initial parameters from the initial data. If not, the initial parameters used are (0, 0, 0).}

\item{tau_mu_log}{whether tau_mu is log-transformed. It should be log-transformed unless set to be exactly zero.}

\item{tau_delta_log}{whether tau_delta is log-transformed. It should be log-transformed unless set to be exactly zero.}

\item{optim_method}{what method to use for the optim function (Nelder-Mead (default), Brent (for 1-dimensional optimization), or L-BFGS-B (not recommended))}

\item{select_method}{The method by which variants were selected: "single_exposure" = a single population's exposure association p-value, "minp" = minimum exposure association p-value across all populations, "fisher" = Fisher's method}

\item{single_exp_pop}{If select_method = "single_exposure", what exposure population to apply it to}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{resample_number}{If select_method = "fisher", the number of resamples performed to estimate the probability of a variant being selected under Fisher's method.}

\item{resample_seed}{The seed used for resampling}

\item{standardized}{A boolean indicating whether summary statistics are standardized; i.e. whether they have the same set of standard errors. If they do, we only need to calculate the probability of selection once across all variants.}

\item{verbose}{A boolean indicating whether to print out current values of parameters at every iteration of optim(). Use for troubleshooting purposes.}
}
\value{
the output from optim(), detailing the optimized values for gamma, tau_mu and tau_delta, information about convergence, and the negative log-likelihood
}
\description{
Optimizing the conditional log-likelihood under variant selection for a list of selected variants.
}
