% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_selection.R
\name{selection_loglik_single}
\alias{selection_loglik_single}
\title{Log-likelihood function under simple model with variant selection (single-variant)}
\usage{
selection_loglik_single(
  sumstat_beta,
  sumstat_se,
  gamma,
  is_overlap = FALSE,
  r_mat = NA,
  tau_mu,
  tau_delta,
  tau_mu_log = FALSE,
  tau_delta_log = FALSE,
  select_method = c("single_exposure", "minp", "fisher"),
  single_exp_pop = 1,
  select_pthresh = 0.05,
  resample_number = 1e+05,
  resample_seed = 2025
)
}
\arguments{
\item{sumstat_beta}{the GWAS effect size estimates for a single variant across K+1 studies: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations.}

\item{sumstat_se}{the standard errors of the effect size estimates in sumstat_beta}

\item{gamma}{The exposure-outcome causal effect parameter}

\item{is_overlap}{a boolean whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat}{a matrix of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap.}

\item{tau_mu}{The variance of the normal distribution underlying mu. It should be log-transformed unless set to be exactly zero.}

\item{tau_delta}{The variance of the normal distributions underlying delta. It should be log-transformed unless set to be exactly zero.}

\item{tau_mu_log}{Whether tau_mu represents its log-transformed value or not}

\item{tau_delta_log}{Whether tau_delta represents its log-transformed value or not}

\item{select_method}{The method by which variants were selected: "single_exposure" = a single population's exposure association p-value, "minp" = minimum exposure association p-value across all populations, "fisher" = Fisher's method}

\item{single_exp_pop}{If select_method = "single_exposure", what exposure population to apply it to}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{resample_number}{If select_method = "fisher", the number of resamples performed to estimate the probability of a variant being selected under Fisher's method.}

\item{resample_seed}{The seed used for resampling}
}
\value{
A log-likelihood for the summary statistics of a single (selected) variant given some parameters gamma, mu, tau
}
\description{
A conditional log-likelihood function for our simple model that accounts for variants being pre-selected using either a single population's exposure association p-value, minimum exposure association p-value across all populations, or Fisher's method.
}
