% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood.R
\name{set_initial_params}
\alias{set_initial_params}
\title{Setting Initial Parameters for Better Optim()}
\usage{
set_initial_params(
  sumstat_beta_list,
  sumstat_se_list,
  tau_mu_log = FALSE,
  tau_delta_log = FALSE
)
}
\arguments{
\item{sumstat_beta_list}{a list of vectors of GWAS effect size estimates with length K+1: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations. Each vector in the list represents summary statistics for one of many variants.}

\item{sumstat_se_list}{a list of vectors of the standard errors for GWAS effect size estimates in sumstat_beta_list}

\item{tau_mu_log}{whether tau_mu is log-transformed. It should be log-transformed unless set to be exactly zero.}

\item{tau_delta_log}{whether tau_delta is log-transformed. It should be log-transformed unless set to be exactly zero.}
}
\value{
A named vector of initial starting parameters for optim()
}
\description{
Numerical optimization methods may be sensitive to choice of initial parameters. This function uses a method of moments-like estimator to set more plausible initial parameters for optimization of the log-likelihood function over gamma, tau_mu and delta
}
\examples{
nvar <- 100
gamma <- 0
h2 <- 0.20
persnp_h2 <- h2 / nvar
tau_mu <- 0 #persnp_h2 to set tau_delta = 0, 0 to set tau_delta = 1
tau_delta <- persnp_h2 - tau_mu
N_k <- c(100000, 10000, 40000, 50000)
sigma_Xk <- 1/sqrt(N_k)
SE_list <- rep(list(sigma_Xk), nvar)

i <- 1
observed_data <- simplemodel_sim(gamma = gamma, tau_mu = tau_mu, tau_delta = tau_delta,
                                 SE_list = SE_list, vars = nvar, pops = 3, seed = i)
sumstat_beta_list <- apply(observed_data$beta_matrix, MARGIN = 1, function(x) {return(x)}, simplify = FALSE)
sumstat_beta_list2 <- lapply(sumstat_beta_list, function(x){return(x[1:4])})
SE_list2 <- lapply(SE_list, function(x){return(x[1:4])})
set_initial_params(sumstat_beta_list = sumstat_beta_list2, sumstat_se_list = SE_list2, tau_mu_log = TRUE, tau_delta_log = TRUE)
}
