% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood.R
\name{simple_loglik_full}
\alias{simple_loglik_full}
\title{Log-likelihood function under simple model (multi-variant)}
\usage{
simple_loglik_full(
  sumstat_beta_list,
  sumstat_se_list,
  gamma,
  is_overlap = FALSE,
  r_mat_list = NA,
  tau_mu,
  tau_delta
)
}
\arguments{
\item{sumstat_beta_list}{a list of vectors of GWAS effect size estimates with length K+1: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations. Each vector in the list represents summary statistics for one of many variants.}

\item{sumstat_se_list}{a list of vectors of the standard errors for GWAS effect size estimates in sumstat_beta_list}

\item{gamma}{The exposure-outcome causal effect parameter}

\item{is_overlap}{a boolean describing whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat_list}{a list of matrices of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap}

\item{tau_mu}{The variance of the normal distribution underlying mu}

\item{tau_delta}{The variance of the normal distributions underlying delta}
}
\value{
A log-likelihood for the summary statistics of a set of variants given some parameters gamma, mu, tau
}
\description{
The marginal log-likelihood for the summary statistics of a list of variants under our meta-analysis model, assuming no selection, and i.i.d. normal mu and delta
}
\examples{
simple_loglik_full(sumstat_beta_list = list(c(0,0,0), c(0,0,0), c(0,0,0)),
                     sumstat_se_list = list(c(1,1,1), c(1,1,1), c(1,1,1)),
                     gamma = 0, tau_mu = 1, tau_delta = 1)
simple_loglik_full(sumstat_beta_list = list(c(0.4, 0.6, 0.3), c(0.7,1,0.2), c(0.3,0.36,0.3)),
                     sumstat_se_list = list(c(0.1, 0.1, 0.05), c(0.2, 0.2, 0.1), c(0.1, 0.1, 0.05)),
                     is_overlap = TRUE,
                     r_mat_list <- rep(list(matrix(c(1, 0.2, 0, 0.2, 1, 0, 0, 0, 1), nrow = 3, ncol = 3)), 3),
                     gamma = 0.8, tau_mu = 1, tau_delta = 1)
}
