% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood.R
\name{simple_loglik_single}
\alias{simple_loglik_single}
\title{Log-likelihood function under simple model (single-variant)}
\usage{
simple_loglik_single(
  sumstat_beta,
  sumstat_se,
  gamma,
  is_overlap = FALSE,
  r_mat = NA,
  tau_mu,
  tau_delta,
  tau_mu_log = FALSE,
  tau_delta_log = FALSE,
  select_zscore = NA
)
}
\arguments{
\item{sumstat_beta}{the GWAS effect size estimates for a single variant across K+1 studies: first, for the outcome in the target population, then the exposure in the target population, then the exposures across K-1 auxiliary populations.}

\item{sumstat_se}{the standard errors of the effect size estimates in sumstat_beta}

\item{gamma}{The exposure-outcome causal effect parameter}

\item{is_overlap}{a boolean whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat}{a matrix of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap.}

\item{tau_mu}{The variance of the normal distribution underlying mu. It should be log-transformed unless set to be exactly zero.}

\item{tau_delta}{The variance of the normal distributions underlying delta. It should be log-transformed unless set to be exactly zero.}

\item{tau_mu_log}{Whether tau_mu represents its log-transformed value or not}

\item{tau_delta_log}{Whether tau_delta represents its log-transformed value or not}

\item{select_zscore}{if the instrument was selected because its association with the exposure in the target exceeded some Z-score threshold, what threshold this is. If no selection was performed to obtain the summary statistics, put NA.}
}
\value{
A log-likelihood for the summary statistics of a single variant given some parameters gamma, mu, tau
}
\description{
The marginal log-likelihood for the summary statistics of a single variant under our meta-analysis model, assuming i.i.d. normal mu and delta. Now includes an option to specify whether variants were selected based on a Z-score threshold for the target exposure.
}
\examples{
simple_loglik_single(sumstat_beta = c(0,0,0), sumstat_se = c(1,1,1),
                     gamma = 0, tau_mu = 1, tau_delta = 1)
simple_loglik_single(sumstat_beta = c(0,0,0), sumstat_se = c(1,1,1),
                     gamma = 0, tau_mu = 1, tau_delta = 0)
simple_loglik_single(sumstat_beta = c(0.4, 0.6, 0.3), sumstat_se = c(0.1, 0.1, 0.05),
                     is_overlap = TRUE,
                     r_mat = matrix(c(1, 0.2, 0, 0.2, 1, 0, 0, 0, 1), nrow = 3, ncol = 3),
                     gamma = 0.8, tau_mu = 1, tau_delta = 1)
simple_loglik_single(sumstat_beta = c(0.4, 0.6, 0.3), sumstat_se = c(0.1, 0.1, 0.05),
                     is_overlap = TRUE,
                     r_mat = matrix(c(1, 0.2, 0, 0.2, 1, 0, 0, 0, 1), nrow = 3, ncol = 3),
                     gamma = 0.8, tau_mu = 0, tau_delta = 0,
                     tau_mu_log = TRUE, tau_delta_log = TRUE)
}
