% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{simplemodel_sim}
\alias{simplemodel_sim}
\title{Simulating summary statistics under our simple hierarchical model}
\usage{
simplemodel_sim(
  gamma,
  tau_mu,
  tau_delta,
  vars,
  pops,
  SE_list,
  r_mat = NA,
  seed = 2025
)
}
\arguments{
\item{gamma}{The exposure-outcome causal effect}

\item{tau_mu}{The variance of the normal distribution underlying mu (central values)}

\item{tau_delta}{The variance of the normal distribution underlying the delta (deviation terms)}

\item{vars}{the number of variants to simulate for}

\item{pops}{the number of populations to simulate data from, consisting of one target population and K-1 auxiliary populations.}

\item{SE_list}{a list of vectors of desired standard errors for each variant}

\item{r_mat}{a matrix of estimated (residual) summary statistic correlations due to sample overlap between each set of GWAS summary statistics. If not provided by default, it will create a matrix assuming no sample overlap.}

\item{seed}{setting a seed for reproducibility}
}
\value{
a list, consisting of the simulated summary statistics in an N x (K + 1) matrix (beta_matrix), the standard errors of each of the summary statistics (se_matrix), the simulated expectations of the summary statistics (mean_matrix), and the underlying input parameters (r_mat, gamma, tau_mu, and tau_delta) for reproducibility
}
\description{
This model generates independent sets of summary statistics for a target population and K auxiliary populations under a simple version of our hierarchical model, where mu and delta terms are normally distributed with variances tau_mu and tau_delta and all are independent of each other. This uses the mvtnorm package.
}
\examples{
SE_list <- rep(list(c(0.4, 0.4, 0.2, 0.1)), 10)
simplemodel_sim(gamma = 0, tau_mu = 1, tau_delta = 1, SE_list = SE_list, vars = 10, pops = 3)
SE_list <- rep(list(c(0.2, 0.1, 0.05, 0.05)), 50)
r_mat <- diag(4)
r_mat[1,2] <- 0.2
r_mat[2,1] <- 0.2
simplemodel_sim(gamma = 0.7, tau_mu = 0.5, tau_delta = 0.2, SE_list = SE_list, vars = 50, pops = 3, r_mat = r_mat)
}
