% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_likelihood_ash.R
\name{metamrash_em_select}
\alias{metamrash_em_select}
\title{Optimizing the conditional log-likelihood under variant selection and ASH pleiotropy for a list of selected variants
Due to the large number of parameters that need to be optimized (gamma, tau_mu, tau_delta, and especially ash_pi), direct numerical optimization via optim() is unreliable. Instead, we opt for an EM algorithm.}
\usage{
metamrash_em_select(
  sumstat_beta_list,
  sumstat_se_list,
  is_overlap = FALSE,
  r_mat_list = NA,
  kernel_matrix = NA,
  sigma_c = c(10^seq(-10, 0, by = 2)),
  closest_exposure_pop = 1,
  select_flag = FALSE,
  select_method = c("single_exposure", "minp", "fisher", "none"),
  matching_exp_pop = "none",
  single_exp_pop = 1,
  select_pthresh = 0.05,
  random_seed = 2026,
  mc_iter = 250,
  fixed_gamma = NA,
  fixed_tau_mu = NA,
  fixed_tau_delta = NA,
  fixed_ash_pi = NA,
  em_iter = 100,
  em_tol = 1e-05,
  standardized = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sumstat_beta_list}{a list of vectors of GWAS effect size estimates with length K+1: first, for the outcome in the target population, then for exposures across K studies. Each vector in the list represents summary statistics for one of J total variants.}

\item{sumstat_se_list}{a list of vectors of the standard errors for GWAS effect size estimates in sumstat_beta_list}

\item{is_overlap}{a boolean describing whether there is any sample overlap between any of the K+1 studies used for summary statistics. Usually this is true due to overlap between the outcome and exposure GWAS in the target population, but we assume no overlap by default.}

\item{r_mat_list}{a list of J matrices of estimated (residual) summary statistics correlations due to sample overlap between each set of GWAS summary statistics. Our method does not provide a matrix by default, but specifies a simple diagonal matrix in the event of no sample overlap}

\item{kernel_matrix}{a K+1 x K+1 correlation matrix that captures correlations in exposure effect size heterogeneity across all populations (including the target population).}

\item{sigma_c}{A vector of length c representing the variance of each of the components of the ASH distribution we used. By default, we specify c = 7 with components 0, 1e-10, 1e-8, 1e-6, 1e-4, 1e-2, and 1.}

\item{closest_exposure_pop}{A single number (from 1 to K) indicating which exposure study is sampled from a population best matches the target population. Used for variant selection.}

\item{select_flag}{A boolean that indicates whether the EM algorithm should account for selection for not. Ignoring selection allows for a closed-form (but possibly inaccurate) algorithm, while accounting for selection needs to be done numerically.}

\item{select_method}{The method by which variants were selected: "single_exposure" = a single population's exposure association p-value, "minp" = minimum exposure association p-value across all populations, "fisher" = Fisher's method.}

\item{matching_exp_pop}{How many exposure studies are sampled from the same population as the target. The EM algorithm changes slightly depending on whether we have population-matched exposures}

\item{single_exp_pop}{If select_method = "single_exposure", what exposure study it applies to. By default, we assume it is the first.}

\item{select_pthresh}{The p-value threshold for variant selection}

\item{random_seed}{For Fisher's method of selection, what random seed to use for sampling of observations}

\item{mc_iter}{For Fisher's method of selection, how many random samples to take}

\item{fixed_gamma}{A single number indicating the value that gamma should be fixed at through the EM algorithm. Set to NA by default to indicate that it should not be fixed.}

\item{fixed_tau_mu}{A single number indicating the value that tau_mu should be fixed at through the EM algorithm. Set to NA by default to indicate that it should not be fixed.}

\item{fixed_tau_delta}{A single number indicating the value that tau_delta should be fixed at through the EM algorithm. Set to NA by default to indicate that it should not be fixed.}

\item{fixed_ash_pi}{A vector of length c indicating the values that ash_pi should be fixed at. Set to NA by default to indicate that it should not be fixed.}

\item{em_iter}{The maximum number of iterations to perform for the EM algorithm.}

\item{em_tol}{The tolerance threshold for stopping the EM algorithm early. If the Euclidean distance between successive EM iterations follows below this threshold, the algorithm ends early.}

\item{standardized}{A boolean indicating whether summary statistics are standardized; i.e. whether they have the same set of standard errors. If they do, we only need to calculate the probability of selection once across all J variants.}

\item{verbose}{If true, prints out the current values of each parameter at each iteration, as well as the observed data log-likelihood.}
}
\value{
A list containing the final parameter values estimated by the EM algorithm, as well as successive iterations. params = the final quantities for gamma/tau_mu/tau_delta/ash_pi, final_loglik = the final observed data log-likelihood at params, params_list = a list of values that each parameter took across successive iterations, loglik_vector = the observed data log-likelihood across all iterations.
}
\description{
Optimizing the conditional log-likelihood under variant selection and ASH pleiotropy for a list of selected variants
Due to the large number of parameters that need to be optimized (gamma, tau_mu, tau_delta, and especially ash_pi), direct numerical optimization via optim() is unreliable. Instead, we opt for an EM algorithm.
}
